+++ docs/scripts/views/components/form.vue
@@ -8,19 +8,21 @@
     <section class="demo-wrapper">
       <div class="demo">
         <ui-form>
-          <legend>Horizontal Form</legend>
-          <ui-form-field class="form-item">
-            <label>Input:</label>
-            <ui-textfield></ui-textfield>
-          </ui-form-field>
-          <ui-form-field class="form-item">
-            <label>Select:</label>
-            <ui-select></ui-select>
-          </ui-form-field>
-          <ui-form-field class="form-action">
-            <ui-button raised>Submit</ui-button>
-            <ui-button outlined>Cancel</ui-button>
-          </ui-form-field>
+          <template #default="{ actionClass }">
+            <legend>Horizontal Form</legend>
+            <ui-form-field>
+              <label>Input:</label>
+              <ui-textfield></ui-textfield>
+            </ui-form-field>
+            <ui-form-field>
+              <label>Select:</label>
+              <ui-select></ui-select>
+            </ui-form-field>
+            <ui-form-field :class="actionClass">
+              <ui-button raised>Submit</ui-button>
+              <ui-button outlined>Cancel</ui-button>
+            </ui-form-field>
+          </template>
         </ui-form>
       </div>
       <ui-snippet :code="$store.demos[1]"></ui-snippet>
@@ -30,15 +32,15 @@
       <div class="demo">
         <ui-form type="|">
           <legend>Vertical Form</legend>
-          <ui-form-field class="form-item">
+          <ui-form-field>
             <label>Input:</label>
             <ui-textfield></ui-textfield>
           </ui-form-field>
-          <ui-form-field class="form-item">
+          <ui-form-field>
             <label>Select:</label>
             <ui-select fullwidth></ui-select>
           </ui-form-field>
-          <ui-form-field class="form-action">
+          <ui-form-field>
             <ui-button raised>Submit</ui-button>
             <ui-button outlined>Cancel</ui-button>
           </ui-form-field>